+++ bk2ch06p313pageControllerBug/ch19p628pageControllerBug/MyPage.m
@@ -9,12 +9,12 @@
 
 - (void) viewDidLoad {
     [super viewDidLoad];
-    self.lab.text = [NSString stringWithFormat:@"%i",self.num];
+    self.lab.text = [NSString stringWithFormat:@"%ld",(long)self.num];
 }
 
 - (void) setNum: (NSInteger) newNum {
     _num = newNum;
-    self.lab.text = [NSString stringWithFormat:@"%i",newNum];
+    self.lab.text = [NSString stringWithFormat:@"%ld",(long)newNum];
 }
 
 @end
